feat(bundles): Point package.json main: to the umd bundle. Stop shipping commonjs.
BREAKING CHANGE:  We no longer ship commonjs in the npm package.   We now ship esm+es5 (individual files) and UMD bundles only.

This brings ui-router-ng2 npm package in line with how the core Angular 2 packages are packaged.

We compile and distribute esm+es5 + typescript typings in `lib/` (formerly in `lib-esm/`)
We *no longer include the commonjs+es5* (formerly in `lib/`)
We now point the `main:` entry to the UMD bundle for commonjs users.